feat: route to KYCC via browser
